vision/requirements.txt
opencv-python>=4.13.0
numpy>=2.0
Pillow>=11.0
requests>=2.32
pytest>=8.0
